chore: Small cleanup and finally tackling materials

+++ Engine/src/core/App.cpp
@@ -74,10 +74,10 @@ void App::run()
 		layer->construct(*this);
 	}
 
-	static constexpr float minimum_time_step = 1000.0F * 0.0333F;
-	static constexpr auto minimum_hertz = 1000.0F / minimum_time_step;
+	static constexpr double minimum_time_step = 1000.0 * 0.0333;
+	static constexpr auto minimum_hertz = 1000.0 / minimum_time_step;
 	static auto current_time = Clock::ms();
-	static float step = minimum_time_step;
+	static double step = minimum_time_step;
 	while (!window->should_close()) {
 		const auto could_prepare = could_prepare_frame();
 
@@ -105,7 +105,7 @@ void App::run()
 		statistics.frame_time = Clock::ms() - current_time;
 		current_time = Clock::ms();
 
-		step = glm::min<float>(statistics.frame_time, minimum_time_step);
+		step = glm::min<double>(statistics.frame_time, minimum_time_step);
 	}
 
 	wait_for_idle(*device);
